tweak how nodes are scored
the search tries to maximize node.len, which used to correspond to the
number of pushes (or more accurately box lines), but now includes
activations of popup walls as well. this leads to the algorithm giving
as much weight to stepping on a popup wall as pushing a block, which
isn't what we want. pushes are more important.

we could say that popup walls don't contribute to node.len at all, but
that has some weird effects, so don't do that. instead score them as
half of a push.  we also have to score a push + popup wall move with the
combined weight of both, so that the manner in which a wall is popped
doesn't affect the search outcome.

here's the score table:
    pull onto a popup wall = 3
    block push = 2
    popup wall = 1

(node.len should probably be renamed to node.score)