activate toggle if we end on it after a pull
i'm not sure if this is right - toggles only activate when you step on
them, not when you step off them (so if we end a pull on one it means we
stepped off it). otoh, if we leave the square on the next turn then we
would have had to step on the button to get into position to push the
block... which means you would be forced to push the button but also
means that you could toggle it as much as you want.